requirements.txt
discord.py==2.6.3
PyYAML==6.0.2
colorama==0.4.6
PyYAML==6.0.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
pip==25.2
propcache==0.3.2
setuptools==65.5.0
typing_extensions==4.14.1
yarl==1.20.1
